refactor(ai): organize provider management modules
